A faster implementation of range unit addresses
Splits the cases within ranges from those not in ranges and handles them separately instead of having the query optimizer have to worry about the disjunction. Takes advantage of the fact that we know there is never going to be a unit that is both within a non-ranged address and within a ranged address -- no duplication.