Use the service_area_summary table to determine service areas
This is a lot faster than doing the geographic joins.

NOTE: This uses SQLAlchemy's schema introspection feature to construct the ServiceAreaSummary model. This introspection is done at model class creation time (i.e., when the application first loads). This means that whenever the schema of that table changes, the application will need to be restarted.